pyproject.toml
[project]
name = "robot-maia"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "faster-whisper>=1.2.1",
    "fastapi>=0.116.1",
    "google-genai>=1.68.0",
    "markdown>=3.10.2",
    "nats-py>=2.14.0",
    "piper-tts>=1.4.1",
    "scipy>=1.17.1",
    "sounddevice>=0.5.5",
    "soundfile>=0.13.1",
    "uvicorn>=0.35.0",
    "zmq>=0.0.0",
    "python-multipart>=0.0.26",
]

[project.scripts]
api = "mia.api:main"
core = "mia.core:main"

[tool.uv]
package = true